Refactor AutoTuner and move it to tuning package
